executor: created reusable pid stats collection utility that can be used in an executor
